Updating import and tests
