+++ universal-login-commons/src/index.ts
@@ -37,7 +37,7 @@ export {normalizeBigNumber, bigNumberifyDecimal} from './core/utils/bigNumbers';
 export {calculateInitializeSignature, calculateInitializeWithENSSignature, getInitializeSigner} from './core/utils/calculateSignature';
 export {clamp} from './core/utils/clamp';
 export {computeContractAddress, computeCounterfactualAddress} from './core/utils/contracts/computeContractAddress';
-export {getContractHash, getDeployedBytecode, isContract} from './core/utils/contracts/contractHelpers';
+export {getContractHash, getDeployedBytecode} from './core/utils/contracts/contractHelpers';
 export {convertTenthGweiToWei} from './core/utils/conversion';
 export {debounce} from './core/utils/debounce';
 export {deepMerge} from './core/utils/deepMerge';
@@ -71,7 +71,6 @@ export {sleep, waitUntil} from './core/utils/wait';
 export {BalanceChecker} from './integration/ethereum/BalanceChecker';
 export {deployContract, deployContractAndWait, DEPLOY_GAS_LIMIT, getDeployData} from './integration/ethereum/deployContract';
 export {withENS} from './integration/ethereum/ens';
-export {fetchHardforkVersion} from './integration/ethereum/fetchHardforkVersion';
 export {resolveName} from './integration/ethereum/resolveName';
 export {stringToWei} from './integration/ethereum/stringToWei';
 export {TokenDetailsService} from './integration/ethereum/TokenDetailsService';
@@ -100,3 +99,4 @@ export {GasPriceOracle} from './integration/ethereum/gasPriceOracle';
 export {GasPriceEstimation, GasPriceOption} from './core/models/GasPriceSuggestion';
 export {fetch} from './integration/http/fetch';
 export {NodeEnv, asNodeEnv} from './core/models/NodeEnv';
+export {ProviderService} from './integration/ethereum/ProviderService';
